Remove use of expect.
Because expect panics, let's avoid using it (hence the lints we added
in previous commits).

This did require reading the whole file at once, because handling any
I/O errors inside the .map call is difficult. I likely could have done
it in a more memory efficient manner, but this is not an embedded
system, our files are small, and our RAM is big. Also, this may
actually be faster because we can just read in the whole file in one
batch of disk accesses.